power: supply: qti: Fix real_type detection with A12 firmware
Xiaomi seems to have modified this enum in their driver as well
as in firmware (presumably adsp) of MIUI 13. Add a dummy element
to bump up the property id to what the firmware expects. Without
this change, real_type always reports as "Unknown", breaking
userspace batterysecret daemon.

Required for fast charging to work with MIUI 13 firmware.

Signed-off-by: Adithya R <gh0strider.2k18.reborn@gmail.com>

[AmeChan]: Apply for odin charger.